adding a helper for tests
